+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/life/ActivityDataRecordResponseService.java
@@ -1,5 +1,7 @@
 package com.dwarfeng.familyhelper.webapi.stack.service.life;
 
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityDataRecordCreateInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityDataRecordUpdateInfo;
 import com.dwarfeng.familyhelper.life.stack.bean.entity.ActivityDataRecord;
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.life.DispActivityDataRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
@@ -29,6 +31,12 @@ public interface ActivityDataRecordResponseService extends Service {
     PagedData<ActivityDataRecord> childForItemRecordedDateDesc(LongIdKey activityDataItemKey, PagingInfo pagingInfo)
             throws ServiceException;
 
+    PagedData<ActivityDataRecord> childForActivityRecordedDateAsc(LongIdKey activityKey, PagingInfo pagingInfo)
+            throws ServiceException;
+
+    PagedData<ActivityDataRecord> childForActivityRecordedDateDesc(LongIdKey activityKey, PagingInfo pagingInfo)
+            throws ServiceException;
+
     DispActivityDataRecord getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException;
 
     PagedData<DispActivityDataRecord> allDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
@@ -40,4 +48,18 @@ public interface ActivityDataRecordResponseService extends Service {
     PagedData<DispActivityDataRecord> childForItemRecordedDateDescDisp(
             StringIdKey accountKey, LongIdKey activityDataItemKey, PagingInfo pagingInfo
     ) throws ServiceException;
+
+    PagedData<DispActivityDataRecord> childForActivityRecordedDateAscDisp(
+            StringIdKey accountKey, LongIdKey activityKey, PagingInfo pagingInfo
+    ) throws ServiceException;
+
+    PagedData<DispActivityDataRecord> childForActivityRecordedDateDescDisp(
+            StringIdKey accountKey, LongIdKey activityKey, PagingInfo pagingInfo
+    ) throws ServiceException;
+
+    LongIdKey create(StringIdKey userKey, ActivityDataRecordCreateInfo createInfo) throws ServiceException;
+
+    void update(StringIdKey userKey, ActivityDataRecordUpdateInfo updateInfo) throws ServiceException;
+
+    void remove(StringIdKey userKey, LongIdKey key) throws ServiceException;
 }